+++ src/application/index.ts
@@ -1,2 +1,2 @@
 export type * from '@/application/types';
-export * from '@/application/ApplicationFactory';
+export * from '@/application/create-application';
